requirements.txt
streamlit
youtube-transcript-api
google-generativeai 
python-dotenv
openai